refactor: make /types directory in utils
