+++ server/src/utils/types/QuestionRequestBody.ts
@@ -1,10 +1,8 @@
-import { Level, Subject } from "./constants";
+import { Level, Subject } from "../constants";
 
-interface QuestionRequestBody {
+export interface QuestionRequestBody {
   title?: string | undefined;
   markdown?: string | undefined;
   level?: Level | undefined;
   subject?: Subject | undefined;
 }
-
-export { QuestionRequestBody };
